feat : fixing token ???

+++ frontend/src/components/pages/Offer/OfferCard.jsx
@@ -30,7 +30,7 @@ const OfferCard = ({
     const [detailed, setDetailed] = useState(false);
     const [isEnable, setEnable] = useState(sug?.dispo);
     const [popupVisible, setPopupVisible] = useState(false);
-    const {timeSince} = useApp();
+    const {timeSince, showConfirmPopup} = useApp();
     const {saveOffer, retireOffer, getSavedOffers, offer} = useOffer();
     const navigate = useNavigate();
     const image = SERVERLINK + "/" + sug?.profileimage;
@@ -39,7 +39,19 @@ const OfferCard = ({
               dolor sit amet Lorem ipsum dolor sit amet Lorem ipsum dolor sit
               amet, Lorem ipsum dolor
               Lorem ipsum dolor sit amet Lorem ipsum dolor sit amet`;
+    const {setMessagePopup} = useAnimation();
 
+    const {
+        getOfferById,
+        getCurrentUserOffers,
+        setUpdateOffer,
+        setUnavalaibleOffer,
+        setAvalaibleOffer,
+        deleteOffer,
+    } = useOffer();
+
+    const {token} = useAuth();
+    
     const contactUser = () => {
         localStorage.setItem(
             "userToChat",
@@ -68,20 +80,7 @@ const OfferCard = ({
     const toggleOfferCardPopup = () => {
         setPopupVisible((prev) => !prev);
     };
-
-    const {setMessagePopup} = useAnimation();
-
-    const {
-        getOfferById,
-        getCurrentUserOffers,
-        setUpdateOffer,
-        setUnavalaibleOffer,
-        setAvalaibleOffer,
-        deleteOffer,
-    } = useOffer();
-
-    const {token} = useAuth();
-
+    
     const NJCAM = {
         Infos: "Njcam est une entrprise qui prend en charge la vente de Camera de securite et la surveillance Lorem ipsum dolor sit amet consectetur adipisicing elit. Eveniet et blanditiis eaque, enim vitae perspiciatis. Porro minus iusto voluptatem, ducimus nemo autem cum, placeat iste nostrum tempora quibusdam quia commodi",
         Image: SERVERLINK + '/njcamlogo.jpg',
@@ -90,13 +89,16 @@ const OfferCard = ({
     }
     const handleDeletePost = async () => {
         // TODO :
+        
+        const userConfirmed = await showConfirmPopup("Voulez vous vraiment supprimer ce message ?")
+        if (!userConfirmed)
+            return
+        
         await deleteOffer(sug.offerid);
         await getCurrentUserOffers();
         localStorage.removeItem("offer");
         setUpdateOffer();
 
-        // - Deleting the post
-        // setPopupVisible(false);
         setMessagePopup("Offre supprimé avec success", TOAST_TYPE.success);
     };
